Moved subprocess to own file
Created a file just to wrap subprocess. If I can get this to
work then I'll never need to mess with subprocess and codacy
again.
