+++ naomi/run_command.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+import subprocess
+
+def run_command(command):
+    output = subprocess.check_output(
+        command,shell=False
+    ).decode('utf-8').strip()
+    return output
